health-of-vegetation
this program is for measuring the health of the vegetation using NDVI. 
The example was used here is an image from Landsat 8. The program reads bands 4 and 5 for measuring the NDVI, then saves the value of in a pixel of a new raster. The new raster can be classified using QGIS or ArcGIS software.